moved old files not needed into v1.0.1
