+++ v1.0.1/sessionManager.ts
